# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clinic Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medical-records:
    get:
      operationId: listMedicalRecords
      tags: [medical-records]
      summary: List medical records
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: doctorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of medical records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalRecord"
    post:
      operationId: createMedicalRecord
      tags: [medical-records]
      summary: Create medical record / visit note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicalRecordBody"
      responses:
        "201":
          description: Medical record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /medical-records/{recordId}:
    get:
      operationId: getMedicalRecord
      tags: [medical-records]
      summary: Get medical record
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
    patch:
      operationId: updateMedicalRecord
      tags: [medical-records]
      summary: Update medical record
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicalRecordBody"
      responses:
        "200":
          description: Updated medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"

  # CLINIC NOTICES
...
  /ultrasound:
    get:
      operationId: listUltrasoundRecords
      tags: [ultrasound]
      summary: List ultrasound records
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [requested, in_progress, completed]
      responses:
        "200":
          description: List of ultrasound records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UltrasoundRecord"
    post:
      operationId: createUltrasoundRecord
      tags: [ultrasound]
      summary: Create ultrasound record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUltrasoundBody"
      responses:
        "201":
          description: Ultrasound record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UltrasoundRecord"
  /ultrasound/{ultrasoundId}:
    get:
      operationId: getUltrasoundRecord
      tags: [ultrasound]
      summary: Get ultrasound record
      parameters:
        - name: ultrasoundId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ultrasound record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UltrasoundRecord"
    patch:
      operationId: updateUltrasoundRecord
      tags: [ultrasound]
      summary: Update ultrasound record (add report)
      parameters:
        - name: ultrasoundId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUltrasoundBody"
      responses:
        "200":
          description: Updated ultrasound record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UltrasoundRecord"
  /ultrasound/{ultrasoundId}/images:
    post:
      operationId: uploadUltrasoundImage
      tags: [ultrasound]
      summary: Upload an ultrasound image file to the server
      parameters:
        - name: ultrasoundId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImageUploadBody"
      responses:
        "201":
          description: Image stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagingAttachment"
  /ultrasound/{ultrasoundId}/images/{attachmentId}:
    get:
      operationId: getUltrasoundImageFile
      tags: [ultrasound]
      summary: Download / view an ultrasound image file (authenticated, audited)
      parameters:
        - name: ultrasoundId
          in: path
          required: true
          schema:
            type: integer
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: download
          in: query
          description: When present, served as an attachment (Content-Disposition) instead of inline.
          schema:
            type: string
      responses:
        "200":
          description: Image bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
    delete:
      operationId: deleteUltrasoundImage
      tags: [ultrasound]
      summary: Delete an ultrasound image file
      parameters:
        - name: ultrasoundId
          in: path
          required: true
          schema:
            type: integer
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Image deleted

  # LABORATORY
...
  /inventory:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: expired
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Inventory item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{itemId}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Soft delete inventory item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item archived
  /inventory/{itemId}/adjust:
    post:
      operationId: adjustInventoryStock
      tags: [inventory]
      summary: Apply a signed stock movement (restock / consume / write-off)
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustStockBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{itemId}/transactions:
    get:
      operationId: listInventoryTransactions
      tags: [inventory]
      summary: Stock movement history for an item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock movement ledger (most recent first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryTransaction"

  # NOTIFICATIONS
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        fullNameAr:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, doctor, nurse, front_desk, xray_staff, lab_staff]
        isActive:
          type: boolean
        isOnShift:
          type: boolean
        phone:
          type: string
          nullable: true
        addressLine:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, username, fullName, role, isActive, isOnShift, createdAt]
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        mrn:
          type: string
        idCardNumber:
          type: string
        fullName:
          type: string
        fullNameAr:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female]
        phone:
          type: string
        address:
          type: string
          nullable: true
        bloodType:
          type: string
          nullable: true
        allergies:
          type: string
          nullable: true
        emergencyContact:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, mrn, idCardNumber, fullName, dateOfBirth, gender, phone, isActive, createdAt]
...
    MedicalRecord:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        appointmentId:
          type: integer
          nullable: true
        patient:
          $ref: "#/components/schemas/Patient"
        doctor:
          $ref: "#/components/schemas/User"
        chiefComplaint:
          type: string
        chiefComplaintAr:
          type: string
          nullable: true
        diagnosis:
          type: string
        diagnosisAr:
          type: string
          nullable: true
        treatment:
          type: string
        treatmentAr:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        vitals:
          $ref: "#/components/schemas/Vitals"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, patientId, doctorId, chiefComplaint, diagnosis, treatment, createdAt]
...
    Vitals:
      type: object
      properties:
        bloodPressure:
          type: string
          nullable: true
        heartRate:
          type: integer
          nullable: true
        temperature:
          type: number
          nullable: true
        weight:
          type: number
          nullable: true
        height:
          type: number
          nullable: true
        oxygenSaturation:
          type: number
          nullable: true
    CreateMedicalRecordBody:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        appointmentId:
          type: integer
        chiefComplaint:
          type: string
        chiefComplaintAr:
          type: string
        diagnosis:
          type: string
        diagnosisAr:
          type: string
        treatment:
          type: string
        treatmentAr:
          type: string
        notes:
          type: string
        vitals:
          $ref: "#/components/schemas/Vitals"
      required: [patientId, doctorId, chiefComplaint, diagnosis, treatment]
    UpdateMedicalRecordBody:
      type: object
      properties:
        chiefComplaint:
          type: string
        chiefComplaintAr:
          type: string
        diagnosis:
          type: string
        diagnosisAr:
          type: string
        treatment:
          type: string
        treatmentAr:
          type: string
        notes:
          type: string
        vitals:
          $ref: "#/components/schemas/Vitals"
...
    MediaImage:
      type: object
      properties:
        url:
          type: string
        fileName:
          type: string
        caption:
          type: string
      required: [url]
    ImagingAttachment:
      type: object
      description: Metadata for a server-stored X-ray / ultrasound image file.
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        caption:
          type: string
          nullable: true
        url:
          type: string
          description: Authenticated download endpoint for the bytes.
        createdAt:
          type: string
          format: date-time
      required: [id, fileName, mimeType, sizeBytes, url, createdAt]
    ImageUploadBody:
      # multipart/form-data. The actual upload is performed client-side via
      # customFetch with a FormData body (binary file under the `file` field) —
      # not the generated hook — so `file` is intentionally left untyped here to
      # avoid an unusable z.instanceof(File) in the backend zod (which has no DOM
      # lib and never validates this body; multer + magic-byte sniffing do).
      type: object
      properties:
        file:
          type: string
          description: The binary image file (PNG/JPEG/WEBP), sent as a multipart field.
        caption:
          type: string
      required: [file]
...
    UltrasoundRecord:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        requestedById:
          type: integer
        performedById:
          type: integer
          nullable: true
        patient:
          $ref: "#/components/schemas/Patient"
        requestedBy:
          $ref: "#/components/schemas/User"
        examType:
          type: string
        bodyPart:
          type: string
        bodyPartAr:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        imageFileName:
          type: string
          nullable: true
        images:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/MediaImage"
        orderGroupId:
          type: string
          nullable: true
        report:
          type: string
          nullable: true
        reportAr:
          type: string
          nullable: true
        status:
          type: string
          enum: [requested, in_progress, completed]
        notes:
          type: string
          nullable: true
        notesAr:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, patientId, requestedById, examType, bodyPart, status, createdAt]
    CreateUltrasoundBody:
      type: object
      properties:
        patientId:
          type: integer
        requestedById:
          type: integer
        examType:
          type: string
        bodyPart:
          type: string
        bodyPartAr:
          type: string
        notes:
          type: string
        notesAr:
          type: string
        orderGroupId:
          type: string
      required: [patientId, requestedById, examType, bodyPart]
    UpdateUltrasoundBody:
      type: object
      properties:
        performedById:
          type: integer
        imageUrl:
          type: string
        imageFileName:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/MediaImage"
        report:
          type: string
        reportAr:
          type: string
        status:
          type: string
          enum: [requested, in_progress, completed]
        notes:
          type: string
        notesAr:
          type: string
        bodyPartAr:
          type: string
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        minimumStock:
          type: integer
        expiryDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, category, quantity, unit, minimumStock, isActive, createdAt]
    CreateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        minimumStock:
          type: integer
        expiryDate:
          type: string
          format: date
        notes:
          type: string
      required: [name, category, quantity, unit, minimumStock]
    UpdateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        minimumStock:
          type: integer
        expiryDate:
          type: string
          format: date
        notes:
          type: string
        isActive:
          type: boolean
    AdjustStockBody:
      type: object
      properties:
        delta:
          type: integer
          description: Signed change. Positive = stock in, negative = stock out. Non-zero.
        reason:
          type: string
          enum: [restock, consumed, expired, adjustment]
        note:
          type: string
      required: [delta, reason]
    InventoryTransaction:
      type: object
      properties:
        id:
          type: string
        delta:
          type: integer
        quantityAfter:
          type: integer
        reason:
          type: string
          enum: [initial, restock, consumed, expired, adjustment]
        note:
          type: string
          nullable: true
        performedById:
          type: integer
          nullable: true
        performedByName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, delta, quantityAfter, reason, createdAt]
...